+++ modules/builtin/src/backend/prompts/confirm.ts
@@ -1,7 +1,6 @@
-import { IO, Prompt } from 'botpress/sdk'
+import { IO, Prompt, PromptConfig } from 'botpress/sdk'
 import * as sdk from 'botpress/sdk'
 import { extractEventCommonArgs } from 'common/action'
-import { PromptConfig } from 'common/typings'
 import yn from 'yn'
 
 import commonFields from './common'
