+++ modules/builtin/src/backend/variables/boolean.ts
@@ -1,5 +1,4 @@
-import { BoxedVarContructor, BoxedVariable } from 'botpress/sdk'
-import { FlowVariableConfig, FlowVariableType } from 'common/typings'
+import { BoxedVarContructor, BoxedVariable, FlowVariableConfig, FlowVariableType } from 'botpress/sdk'
 import yn from 'yn'
 
 class BoxedBoolean implements BoxedVariable<boolean> {
