+++ modules/builtin/src/backend/variables/date.ts
@@ -1,5 +1,4 @@
-import { BoxedVarContructor, BoxedVariable } from 'botpress/sdk'
-import { FlowVariableConfig, FlowVariableType } from 'common/typings'
+import { BoxedVarContructor, BoxedVariable, FlowVariableConfig, FlowVariableType } from 'botpress/sdk'
 import moment from 'moment'
 
 type BoxedDateType = string | Date | moment.Moment
