+++ modules/qna/src/views/full/Components/ContentAnswerForm/index.tsx
@@ -1,5 +1,5 @@
 import { Tab, Tabs } from '@blueprintjs/core'
-import { FormData } from 'botpress/common/typings'
+import { FormData } from 'botpress/sdk'
 import { ContentForms, Dropdown, lang, MoreOptions, MoreOptionsItems, RightSidebar } from 'botpress/shared'
 import cx from 'classnames'
 import React, { FC, Fragment, useEffect, useReducer, useRef, useState } from 'react'
