+++ modules/qna/src/views/full/utils/qnaList.utils.ts
@@ -1,4 +1,4 @@
-import { FormData } from 'botpress/common/typings'
+import { FormData } from 'botpress/sdk'
 import { lang } from 'botpress/shared'
 import _ from 'lodash'
 import _uniqueId from 'lodash/uniqueId'
