+++ src/bp/common/typings.ts
@@ -258,82 +258,3 @@ export interface ActionParameterDefinition {
 export type ActionServerWithActions = ActionServer & {
   actions: ActionDefinition[] | undefined
 }
-
-export interface FlowVariableType {
-  id: string
-  config: FlowVariableConfig
-  box: BoxedVarConstructable<any>
-}
-
-export type FlowVariableConfig = FormDefinition
-
-export interface FormData {
-  id?: string
-  contentType?: string
-  [key: string]: undefined | number | boolean | string | FormData[]
-}
-
-export interface FormMoreInfo {
-  label: string
-  url?: string
-}
-
-export interface FormAdvancedSetting {
-  key: string
-  label: string
-  type: string
-  moreInfo?: FormMoreInfo
-}
-
-export interface FormOption {
-  value: string
-  label: string
-  related: FormField
-}
-
-export interface FormContextMenu {
-  type: string
-  label: string
-}
-
-export interface FormField {
-  type: 'checkbox' | 'group' | 'select' | 'text' | 'textarea' | 'upload' | 'url'
-  key: string
-  label: string
-  placeholder?: string
-  options?: FormOption[]
-  fields?: FormField[]
-  group?: {
-    addLabel?: string // you have to specify the add button label
-    minimum?: number // you can specify a minimum so the delete button won't show if there isn't more than the minimum
-    contextMenu?: FormContextMenu[] // you can add a contextual menu to add extra options
-  }
-}
-
-export interface PromptDefinition {
-  id: string
-  config: PromptConfig
-  prompt: PromptConstructable
-}
-
-/** The configuration of the prompt which is saved on the flow */
-export type PromptConfig = {
-  /** An ID used internally to refer to this prompt */
-  type: string
-  /** The label displayed in the studio */
-  label?: string
-  icon?: string
-  /** The ID representing the type of value that is collected by this prompt */
-  valueType?: string
-  /** A list of ID represented by the type of values collected by this prompt */
-  valueTypes?: string[]
-  /** The minimum confidence required for the value to be considered valid */
-  minConfidence?: number
-  /** Whatever happens, the prompt will never ask the user to validate the provided value */
-  noValidation?: boolean
-} & FormDefinition
-
-export interface FormDefinition {
-  advancedSettings: FormAdvancedSetting[]
-  fields: FormField[]
-}
